Background addition
Added background, tried to configure VS proj settings so you dont have to. Download sfml and install it such that its contents are in C:/Program Files(x86)/SFML and you'll be fine without any configs.
Will change all the sprites etc. later. This one is the skeleton and filled it with this stuff as a demo ig