added luck factor and bot sentience
BOT SENTIENCE

function Player::decideTurn() is where if the Player is a bot, the bot decides the turn based on the conditions it is currently facing. The factors contributing to its decision are clear in the bool exp.

LUCK FACTOR

the game keeps track of the outcomes of the last three turns of each player. If the player is too lucky, their dice roll is adjusted accordingly. The formula for adjusting is clear (sum of last three differences multiplied by 1.5)

MISC CHANGES

added turnChoice enum for readability

mBaseAttr is now initialized inside constructors (so seed for each player changes)

thievery is removed

changed function Player::damageDealt() such that mHP decreases before if statement

mHP assignment in setAllSkills has changed

NOTE: There are many other changes I may have not mentioned here... so it's best that you simple copy paste the code as they all coordinate with each other! Merci beaucoup